[UNSTABLE] Partition notify-client
