+++ pkg/virt-launcher-libvirt-qemu/notify-client/BUILD.bazel
@@ -1,0 +1,62 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["client.go"],
+    importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/notify-client",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/apimachinery/wait:go_default_library",
+        "//pkg/ephemeral-disk-utils:go_default_library",
+        "//pkg/handler-launcher-com:go_default_library",
+        "//pkg/handler-launcher-com/notify/info:go_default_library",
+        "//pkg/handler-launcher-com/notify/v1:go_default_library",
+        "//pkg/util/net/grpc:go_default_library",
+        "//pkg/virt-launcher/metadata:go_default_library",
+        "//pkg/virt-launcher/virtwrap/agent-poller:go_default_library",
+        "//pkg/virt-launcher/virtwrap/api:go_default_library",
+        "//pkg/virt-launcher/virtwrap/cli:go_default_library",
+        "//pkg/virt-launcher/virtwrap/converter:go_default_library",
+        "//pkg/virt-launcher/virtwrap/errors:go_default_library",
+        "//pkg/virt-launcher/virtwrap/util:go_default_library",
+        "//staging/src/kubevirt.io/api/core/v1:go_default_library",
+        "//staging/src/kubevirt.io/client-go/log:go_default_library",
+        "//vendor/k8s.io/api/core/v1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/util/json:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/watch:go_default_library",
+        "//vendor/k8s.io/client-go/tools/reference:go_default_library",
+        "//vendor/libvirt.org/go/libvirt:go_default_library",
+        "@org_golang_google_grpc//:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = [
+        "client_suite_test.go",
+        "notify_test.go",
+    ],
+    embed = [":go_default_library"],
+    deps = [
+        "//pkg/handler-launcher-com/notify/info:go_default_library",
+        "//pkg/testutils:go_default_library",
+        "//pkg/virt-handler/notify-server:go_default_library",
+        "//pkg/virt-launcher/metadata:go_default_library",
+        "//pkg/virt-launcher/virtwrap/api:go_default_library",
+        "//pkg/virt-launcher/virtwrap/testing:go_default_library",
+        "//pkg/virt-launcher/virtwrap/util:go_default_library",
+        "//staging/src/kubevirt.io/api/core/v1:go_default_library",
+        "//staging/src/kubevirt.io/client-go/api:go_default_library",
+        "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
+        "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
+        "//vendor/github.com/onsi/gomega:go_default_library",
+        "//vendor/go.uber.org/mock/gomock:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/api/equality:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/watch:go_default_library",
+        "//vendor/k8s.io/client-go/tools/cache:go_default_library",
+        "//vendor/k8s.io/client-go/tools/record:go_default_library",
+        "//vendor/libvirt.org/go/libvirt:go_default_library",
+    ],
+)
