+++ pkg/virt-launcher-libvirt-qemu/notify-client/client_suite_test.go
@@ -1,0 +1,11 @@
+package eventsclient
+
+import (
+	"testing"
+
+	"kubevirt.io/client-go/testutils"
+)
+
+func TestClient(t *testing.T) {
+	testutils.KubeVirtTestSuiteSetup(t)
+}
